Add "Save Safari Session URLs"
This script reads URLs from Safari session files (LastSession.plist) and
prints them to standard output (when run from the command line) or saves
them to a text file (otherwise).

AppleScript might seem like an odd choice for this task, and it is. The
original idea was to have Safari actually open the URLs, but it turned
out I never actually need to do that. I usually don't need to look at
a list of URLs either, but I've occasionally found it comforting to
compare backed-up sessions to verify that Safari isn't forgetting tabs.