Add nonmarkup README, move Markdown one to .github
Although GitHub-hosted projects' READMEs are (I expect) primarily read
on the website, I insist on providing one that is reasonably legible in
raw form.  Unfortunately, the Markdown README isn't great for this.

Address this by creating a nonmarkup README that prioritizes legibility.
It's basically the same as the Markdown file, except it has citations in
place of links, the references don't have link titles, and the `compile`
script is not cited (which wouldn't make sense).

(Since I'm already moving the Markdown file, might as well give it the
shorter ".md" extension.)